+++ main/Flap.cpp
@@ -114,7 +114,7 @@ void Flap::prepLevels()
     {
         // adapt speeds to actual wingload
         for ( FlapLevel &fl : flevel ) {
-            fl.prep_speed = fl.nvs_speed * sqrt( (ballast.get()+100.0) / 100.0 );
+            fl.prep_speed = fl.nvs_speed * std::sqrtf( (ballast.get()+100.0) / 100.0 );
             ESP_LOGI( FNAME, "Adjusted flap speed %.1f", fl.prep_speed );
         }
 
@@ -218,7 +218,7 @@ float Flap::getOptimum(float spd) const {
     if (g_force < 0.3) {
         g_force = 0.3; // Ignore meaningless values below 0.3g
     }
-    float g_speed = spd / sqrt(g_force); // reduce current speed, instead of increase switch points
+    float g_speed = spd / sqrtf(g_force); // reduce current speed, instead of increase switch points
     // ESP_LOGI(FNAME, "g force: %.1f, g corrected speed: %3.1f", g_force, g_speed);
 
     int wki = 0; // find the wk index one index above the current speed
