+++ main/KalmanMPU6050.h
@@ -186,12 +186,12 @@ private:
 
   static float fallbackToGyro();
   // static void RollPitchFromAccel(double *roll, double *pitch);
-  static double PitchFromAccel();
-  static double PitchFromAccelRad();
+  static float PitchFromAccel();
+  static float PitchFromAccelRad();
   static uint64_t last_rts;
   static float  filterPitch_rad;
   static float  filterRoll_rad;
-  static double  filterYaw;
+  static float  filterYaw;
 
   static float   fused_yaw;
